refactor: fix typo in FactoryPlan(n)er
