+++ src/components/factory/FactoryPlanner.svelte
